requirements.txt
aiofiles==23.2.1
aiogram==3.8.0  
aiohttp==3.9.5  
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.1.2
frozenlist==1.4.1

pydantic_settings==2.3.4
openai==1.35.10
pyautogen==0.4.6
pypdf==4.2.0